moved HOC into components/hoc folder
