+++ src/components/BarChart.js
@@ -1,6 +1,6 @@
 import createBarChart from './d3/createBarChart'
-import withResponsiveness from '../higherOrderComponents/withResponsiveness'
-import createChartComponent from '../higherOrderComponents/createChartComponent'
+import withResponsiveness from './hoc/withResponsiveness'
+import createChartComponent from './hoc/createChartComponent'
 import compose from '../util/compose'
 
 const BarChart = compose(
