+++ src/components/GradesDistribution.js
@@ -1,6 +1,6 @@
 import createHistogram from './d3/createHistogram'
-import withResponsiveness from '../higherOrderComponents/withResponsiveness'
-import createChartComponent from '../higherOrderComponents/createChartComponent'
+import withResponsiveness from '../components/hoc/withResponsiveness'
+import createChartComponent from '../components/hoc/createChartComponent'
 import compose from '../util/compose'
 
 const Histogram = createChartComponent(createHistogram)
