+++ src/components/Histogram.js
@@ -1,6 +1,6 @@
 import createHistogram from './d3/createHistogram'
-import withResponsiveness from '../higherOrderComponents/withResponsiveness'
-import createChartComponent from '../higherOrderComponents/createChartComponent'
+import withResponsiveness from './hoc/withResponsiveness'
+import createChartComponent from './hoc/createChartComponent'
 import compose from '../util/compose'
 
 export default compose(
