+++ src/components/MapChart.js
@@ -1,6 +1,6 @@
 import createMapChart from './d3/createMapChart'
-import withResponsiveness from '../higherOrderComponents/withResponsiveness'
-import createChartComponent from '../higherOrderComponents/createChartComponent'
+import withResponsiveness from '../components/hoc/withResponsiveness'
+import createChartComponent from '../components/hoc/createChartComponent'
 import compose from '../util/compose'
 
 export default compose(
