+++ src/components/Sankey.js
@@ -1,6 +1,6 @@
 import createSankeyDiagram from './d3/createSankeyDiagram'
-import withResponsiveness from '../higherOrderComponents/withResponsiveness'
-import createChartComponent from '../higherOrderComponents/createChartComponent'
+import withResponsiveness from './hoc/withResponsiveness'
+import createChartComponent from './hoc/createChartComponent'
 import compose from '../util/compose'
 
 export default compose(
