+++ src/components/hoc/createChartCard.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Paper from '@material-ui/core/Paper'
 import Grid from '@material-ui/core/Grid'
-import emojiEndpoints from '../constants/emojiEndpoints'
+import emojiEndpoints from '../../constants/emojiEndpoints'
 
 const createChartCard = (ChartComponent, EmojiFeedback) => props => {
   const {
